Moved json schema parsing to a separated service
